- jira bug fix, updated beneficiary api integration and updated request to pay api integration

+++ lib/feature/payment/send_money/send_money_page_view.dart
@@ -169,7 +169,7 @@ class SendMoneyPageView extends BasePageViewWidget<SendMoneyViewModel> {
                   textColor: Colors.black,
                   rightButtonFn: () {
                     if (int.parse(model.currentPinValue) >
-                        int.parse(ProviderScope.containerOf(context)
+                        double.parse(ProviderScope.containerOf(context)
                             .read(appHomeViewModelProvider)
                             .dashboardDataContent
                             .account!
